engine/0 drawViewFinderHud: the grid-loop bound is a float literal, not a double
The `for (; gridX < (f64)640.0f; ...)` bound made MWCC intern an 8-byte
double and load it with `lfd`.  Retail loads `lbl_803E1F4C` with `lfs`,
and `config/GSAE01/symbols.txt` types that symbol `size:0x4 align:4
data:float`.  The cast was ours, not the original's -- `gridX` is an f32
and the compare never needed double precision.

drawViewFinderHud 99.18072 -> 99.34538 (the residual is now a pure #82
FPR permutation plus one surplus frsp).
engine/0 .sdata2 96.2733 -> 97.2860, unit 99.75118 -> 99.76209.
tree 99.815420 -> 99.815690; matched_code 2517640, matched_data 1204625
and complete_units 921 all HELD.
0 REGRESSED / 0 LOST / 0 UNSCORED / 0 NEW-UNSCORED / 0 DEMOTED, 1 IMPROVED.
ninja all_source EXIT=0; forced main.dol: OK twice, md5 7b955850 == retail.
obj_equal --tree against a control built by reverse-applying this commit
at the same parent: 1013 compared, only engine/0 differs (.text, .sdata2,
.comment), 0 renumbered / 0 missing / 0 new.

POOL, stated in full because score_delta_gate flags it: engine/0 .sdata2
positional hits 94 -> 64 and the section shrinks 952 -> 936 bytes, because
deleting the 8-byte literal shifts the tail by 16.  On the VALUE axis the
move is strictly toward retail -- the multiset delta is `extra
{0x40840000: 1} -> {}` (that word is the high half of the double, and
C66's ledger already recorded it as engine/0's one spurious pool word),
and the three newly "missing" words are all `0x00000000`, the alignment
and low-half padding that only existed because our literal was the wrong
width.  The set of real constants we still lack is unchanged at eight --
0x40490fdb, 0x42a00000, 0x43a00000, 0x43d9999a, 0x43800000, 0x3f700000,
0xc0066666, 0x3a800000 -- the same eight before and after.  Closing that
shortfall is the const lane's row, not this one.

Tree-wide sweep behind the fix: all 210 sub-100 functions in the 87
paired units, retail vs ours mnemonic streams aligned, every hunk trading
a single-precision opcode for its double twin flagged.  This was the only
lfs<->lfd row in the tree.

+++ src/dlls/engine/0/0.c
@@ -1755,7 +1755,7 @@ void drawViewFinderHud(void)
             gridSpacing = lbl_803E1EC4;
             angleDivisor = gGameUiAngleDivisor;
             waveBaseOffset = 479.5;
-            for (; gridX < (f64)640.0f; gridX += gridSpacing)
+            for (; gridX < 640.0f; gridX += gridSpacing)
             {
                 {
                     f32 cosine;
